requirements-dev.txt
bumpversion
ruff
coverage
pytest-cov
